STYLE: Relocate DCMQI headers into "include/dcmqi". See #119
This will streamline use of DCMQI headers in other project by ensuring
the DCMQI headers are included and not the one from project providing
headers with similar names.

After moving the header into "include/dcmqi" directory, the following
one-liner was used to update the files:

  $ cd dcmqi/include/dcmqi
  $ for i in $(ls -1); do (cd ../../; for f in $(ack-grep -l "\"$i"); do sed -i -e "s/\"$i/\"dcmqi\/$i/" $f; done); done
