+++ apps/paramaps/itkimage2paramap.cxx
@@ -2,7 +2,7 @@
 #include "itkimage2paramapCLP.h"
 
 // DCMQI includes
-#include "ParaMapConverter.h"
+#include "dcmqi/ParaMapConverter.h"
 
 int main(int argc, char *argv[])
 {
