+++ apps/seg/segimage2itkimage.cxx
@@ -2,7 +2,7 @@
 #include "segimage2itkimageCLP.h"
 
 // DCMQI includes
-#include "ImageSEGConverter.h"
+#include "dcmqi/ImageSEGConverter.h"
 
 int main(int argc, char *argv[])
 {
