+++ include/dcmqi/Exceptions.h
