+++ include/dcmqi/QIICRConstants.h
